fixed major errors and added a dropdown (work in progress) for ingredient measurement
