--- cake-and-bake/src/app/create-recipe/create-recipe.component.html
@@ -1,32 +1,0 @@
-<div class="wrapper">
-    <h1>Create a new recipe</h1>
-    <div class="row">
-        <form (submit)="createRecipe">
-            <div class="row">
-                <div class="input-field col s12">
-                    <input id="name" name="name" [(ngModel)]="newRecipe.name" required type="text" class="validate">
-                    <label for="name">Name</label>
-                </div>
-                <div class="input-field col s12">
-                    <textarea id="description" [(ngModel)]="newRecipe.description" class="materialize-textarea"></textarea>
-                    <label for="description">Description</label>
-                </div>
-            </div>
-            <button class="btn waves-effect waves-light light-green" type="submit">Submit</button>
-        </form>
-        <div id="myDIV" class="header">
-            <h2>My To Do List</h2>
-            <input type="text" id="myInput" placeholder="Title...">
-            <span onclick="newElement()" class="addBtn">Add</span>
-        </div>
-        
-        <ul id="myUL">
-            <li>Hit the gym</li>
-            <li class="checked">Pay bills</li>
-            <li>Meet George</li>
-            <li>Buy eggs</li>
-            <li>Read a book</li>
-            <li>Organize office</li>
-        </ul>
-    </div>
-</div>
